requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
anyio==3.7.0
    # via httpcore
asgiref==3.7.2
    # via django
black==23.3.0
    # via -r requirements.in
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
click==8.1.3
    # via
    #   black
    #   uvicorn
django==4.2.1
    # via
    #   -r requirements.in
    #   django-admin-sortable2
    #   django-ninja
django-admin-sortable2==2.1.8
    # via -r requirements.in
django-environ==0.10.0
    # via -r requirements.in
django-ninja==0.22.1
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.2
    # via httpx
httpx==0.24.1
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
mypy-extensions==1.0.0
    # via black
packaging==23.1
    # via black
pathspec==0.11.1
    # via black
platformdirs==3.5.1
    # via black
pydantic==1.10.8
    # via django-ninja
sentry-sdk==1.24.0
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlparse==0.4.4
    # via django
typing-extensions==4.6.2
    # via pydantic
urllib3==1.26.16
    # via sentry-sdk
uvicorn==0.22.0
    # via -r requirements.in
whitenoise==6.4.0
    # via -r requirements.in
